requirements.txt
fastapi[all]
pymongo 
uvicorn
python-dotenv
python-openstackclient
python-ldap
pycln
black
pyjwt